Install a hardware-specific wireplumber config in /run/wireplumber
This is to allow different behaviour (bluetooth profile, sink and
source selection, etc) on different devices.

If the manufacturer and model do not have a defined configuration
a default config is installed.

This setup requires either a wireplumber daemon compiled to look
in /run rather than /etc or for /etc/wireplumber to be a symlink
to its /run counterpart.