Enable automatic filter policy in pipewire 1.0+
Smart filters policy allows automatic filter linking tied to a
specific target node.

This is done transparently to the client, allowing users to define a
specific processing chain for a specific device without having to
create setups with virtual sinks (or sources) that must be explicitly
targeted by the clients.