Do not restore filter stream volumes on start up
As these are meant to be hidden the UI cannot be used to check
or set their volumes: This means that if they do get muted somehow
then audio volume will be zero for no readily apparent reason.

Not restoring their saved volume(s) means that audio output will
resume on restart even if something managed to mute them.