wireplumber: Disable the auto profile switch
Automatically switching to HSP/HFP when the microphone is in use is
currently working, but is not yet ready to be enabled by default.

One outstanding issue is that during the codec transition, the audio
output is briefly redirected to the built-in speakers.

The problems with the automatic profile switch are currently tracked in
https://gitlab.steamos.cloud/holo-team/tasks/-/issues/881.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>