wireplumber: Improve the PlayStation controller matches rules
Newer firmware for the DualSense controller changed the name from
"Wireless Controller" to a more specific "DualSense Wireless
Controller".

Given that we can't really rely on the card name field to remain at a
fixed string, we change the matches rule to use the PID and VID instead.

With this commit we add the DualSense, DualSense Edge, and the two
DualShock 4 variants.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>